refactor(web/listings): rename PascalCase component files to kebab-case
Renames the last 10 PascalCase-named component files (plus 7 co-located
test files) in features/listings/components/ to kebab-case, per
docs/frontend-architecture.md, and updates every import/export/mock
site referencing the old filenames. No behavioral change.

Closes #1805
