+++ apps/web/src/features/listings/components/woocommerce-publish-wizard.test.tsx
@@ -11,7 +11,7 @@
 import { screen, fireEvent, waitFor, cleanup } from '@testing-library/react';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { renderWithProviders, createMockApiClient } from '../../../test/test-utils';
-import { WoocommercePublishWizard } from './WoocommercePublishWizard';
+import { WoocommercePublishWizard } from './woocommerce-publish-wizard';
 import type { Connection } from '../../connections';
 
 const wooConnection: Connection = {
