+++ apps/web/src/pages/listings/listing-detail-page.tsx
@@ -8,10 +8,10 @@ import { RawPayloadPanel } from '../../shared/ui/raw-payload-panel';
 import { TimeDisplay } from '../../shared/ui/time-display';
 import { usePlatform } from '../../shared/plugins';
 import { useListingQuery } from '../../features/listings/hooks/use-listing-query';
-import { EditOfferDrawer } from '../../features/listings/components/EditOfferDrawer';
+import { EditOfferDrawer } from '../../features/listings/components/edit-offer-drawer';
 import { ListingMarketplaceOfferSection } from '../../features/listings/components/listing-marketplace-offer-section';
-import { OfferCreationStatusBadge } from '../../features/listings/components/OfferCreationStatusBadge';
-import { OfferCreationErrorList } from '../../features/listings/components/OfferCreationErrorList';
+import { OfferCreationStatusBadge } from '../../features/listings/components/offer-creation-status-badge';
+import { OfferCreationErrorList } from '../../features/listings/components/offer-creation-error-list';
 import { ConnectionEntityLabel } from '../../features/connections/components/ConnectionEntityLabel';
 import { useVariantQuery } from '../../features/products/hooks/use-variant-query';
 import type { ProductVariantSummary } from '../../features/products/api/products.types';
